+++ config/webpack.config.js
@@ -88,7 +88,15 @@ module.exports = ( env, argv ) => {
 								}
 							}
 						},
-						`sass-loader`
+						{
+							loader: `sass-loader`,
+							options: {
+								api: `modern`,
+								sassOptions: {
+									silenceDeprecations: [`mixed-decls`]
+								}
+							}
+						}
 					]
 				}
 			]